requirements.txt
# NeuroLearn-GEN Requirements
# Python 3.11+

# Core LLM & Agents
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
